+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## ChangeLog
 
+### 1.1.0
+
+* Added support to blur, paste, change events
+* Add Human readable parameter
+* Code cleaned
+
 ### 1.0.0
 
 * Initial release